Some gruesome hacks to make static site build work
Ok, so the style.css referred to in reveal-md.json are copied into an
assets directory. It refers to an image contained in a directory, and
this has to be included as a "static-dir" to be included when building a
static site, but the _directory_ it is contained in is not included,
only the contents, which makes the paths different when running a local
site when developing vs. the built static site. It's a mess.